+++ lib/openapi_parser/spec_validator.rb
@@ -27,6 +27,8 @@ require_relative 'spec_validator/rules/security_scheme_fields_before_32'
 require_relative 'spec_validator/rules/media_types_before_32'
 require_relative 'spec_validator/rules/streaming_fields_before_32'
 require_relative 'spec_validator/rules/default_mapping_before_32'
+require_relative 'spec_validator/rules/query_method_before_32'
+require_relative 'spec_validator/rules/additional_operations_before_32'
 
 module OpenAPIParser
   class SpecViolationError < OpenAPIError
@@ -101,6 +103,8 @@ module OpenAPIParser
           Rules::MediaTypesBefore32,
           Rules::StreamingFieldsBefore32,
           Rules::DefaultMappingBefore32,
+          Rules::QueryMethodBefore32,
+          Rules::AdditionalOperationsBefore32,
         ]
       end
   end